Adapted the code for the 3.x branch
