+++ apscheduler/executors/asyncio.py
@@ -6,10 +6,8 @@ from apscheduler.executors.base import BaseExecutor, run_job
 from apscheduler.util import iscoroutinefunction_partial
 
 try:
-    from asyncio import iscoroutinefunction
     from apscheduler.executors.base_py3 import run_coroutine_job
 except ImportError:
-    from trollius import iscoroutinefunction
     run_coroutine_job = None
